+++ www/Library/admin_widgets/page_widgets/PageDescriptionEditor/PageDescriptionEditor.php
@@ -32,6 +32,7 @@ class PageDescriptionEditor extends Widget {
         }
         $v['object'] = $this->_input['REQUEST']['object']->uri();
         $v['value'] = $this->_input['REQUEST']['object']->value();
+        $v['title'] = $this->_input['REQUEST']['object']->title->value();
         return parent::work($v);
     }