+++ www/Library/admin_widgets/page_widgets/PageDescriptionEditor/res/boolive.PageDescriptionEditor/boolive.PageDescriptionEditor.js
@@ -12,38 +12,14 @@
         _create: function() {
             $.boolive.Widget.prototype._create.call(this);
             var self = this;
-            if(self._form==null){
-               self._form = $('<form method="POST" action=""></form>');
-               var input = $('<textarea class="description" name="Page[description]" ></textarea>');
-               var obj = $('<input type="hidden" name="object" value="">');
-               var error = $('<div class="error"></div>');
-            }
-            var a = this.element.find('.descriptionlink');
-            self._value = a.text();
-            a.click(function(e){
-                if(!self._form.hasClass('hide')){
-                   e.preventDefault();
-                   $(input).text($(this).text());
-                   $(obj).val(self.options.object);
-                   $(self._form).append(obj);
-                   $(self._form).append(input);
-                   $(self._form).append(error);
-                   $(this).parent().append(self._form);
-                   $(this).addClass('hide');
-
-                }else{
-                    e.preventDefault();
-                    $(this).addClass('hide');
-                    self._form.removeClass('hide');
-                }
-            });
+            self._form = self.element.find('.form_description');
+            self._value = self._form.find('.description').text();
             self._form.on('keyup', '.description', function(e) {
                 if(self._form.find('.description').val()!=''){
                     self._change(e);
                 }
 
             })
-
         },
         _change: function(e){
             if (this._value != this.element.html()){
@@ -69,19 +45,13 @@
                         }
                     }else{
                         self._form.find('div.error').text('');
-                        self.element.find('a.descriptionlink').text(responseText.out.description);
-                        self.element.find('a.descriptionlink').removeClass('hide');
-                        self._form.addClass('hide');
-                       // self._value = responseText.out.description;
+                        self.element.find('description').val(responseText.out.description);
                     }
                 }
             });
         },
         call_cancel: function(e){
-            if(!this._form.hasClass('hide') &&  this.element.has('form').get(0)){
-                this._form.addClass('hide');
-            }
-            this.element.find('a.descriptionlink').removeClass('hide').text(this._value);
+            this._form.find('.description').val(this._value);
         }
     });
 })(jQuery);