+++ www/config.stores.php
@@ -14,7 +14,7 @@ $config = array(
                 // Тип СУБД
                 'driver' => 'mysql',
                 // Имя базы данных
-                'dbname' => 'boolive2-git-test',
+                'dbname' => 'boolive2',
                 // Адрес сервера
                 'host' => 'localhost',
                 // Порт